// frontend/src/types/proto-es/store/policy_pb.d.ts
// @generated by protoc-gen-es v2.12.0
// @generated from file store/policy.proto (package laelia.store, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";
import type { Expr } from "../google/type/expr_pb";

/**
 * Describes the file store/policy.proto.
 */
export declare const file_store_policy: GenFile;

/**
 * @generated from message laelia.store.Binding
 */
export declare type Binding = Message<"laelia.store.Binding"> & {
  /**
   * The role that is assigned to the members.
   * Format: roles/{role}
   *
   * @generated from field: string role = 1;
   */
  role: string;

  /**
   * Specifies the principals requesting access for a resource.
   * For users, the member should be: users/{userUID}
   * For groups, the member should be: groups/{email}
   *
   * @generated from field: repeated string members = 2;
   */
  members: string[];

  /**
   * The condition that is associated with this binding.
   * If the condition evaluates to true, then this binding applies to the current request.
   * If the condition evaluates to false, then this binding does not apply to the current request. However, a different role binding might grant the same role to one or more of the principals in this binding.
   *
   * @generated from field: google.type.Expr condition = 3;
   */
  condition?: Expr | undefined;
};

/**
 * Describes the message laelia.store.Binding.
 * Use `create(BindingSchema)` to create a new message.
 */
export declare const BindingSchema: GenMessage<Binding>;

/**
 * @generated from message laelia.store.IamPolicy
 */
export declare type IamPolicy = Message<"laelia.store.IamPolicy"> & {
  /**
   * Collection of binding.
   * A binding binds one or more members or groups to a single role.
   *
   * @generated from field: repeated laelia.store.Binding bindings = 1;
   */
  bindings: Binding[];
};

/**
 * Describes the message laelia.store.IamPolicy.
 * Use `create(IamPolicySchema)` to create a new message.
 */
export declare const IamPolicySchema: GenMessage<IamPolicy>;